server/requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.14.6
cryptography==3.4.8
flask
ffmpeg
Fabric3==1.14.post1
google-cloud-speech
inflect==5.3.0
Jinja2==3.0.1
jinja2-pluralize==0.3.0
MarkupSafe==2.0.1
opencv-python
opensmile==2.4.1
paramiko==2.7.2
pycparser==2.20
pydub
PyMySQL==1.0.2
PyNaCl==1.4.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
django-widget-tweaks==1.4.8
librosa
numpy
pandas
soundfile==0.9.0
wave
scikit-learn==0.24.2
tqdm==4.28.1
matplotlib==2.2.3
pyaudio==0.2.11
tensorflow
pyyaml
text_unidecode
torch
torchvision
torchaudio
transformers
emoji
firebase-admin
firebase
vaderSentiment
flask-cors